rtw88-usb: fix the wrong check of fw fifo dump ability
In rtw_fw_dump_fifo(), the code checks if the chip supports dump
fw fifo by using rtwdev->chip->fw_fifo_addr. Meanwhile, fw_fifo_addr
is array not pointer, so it is not NULL always. Propose to check the
fw_fifo_addr[0] instead of fw_fifo_addr, and for the chip which not
support dump fw fifo, we can declare it as .fw_fifo_addr = {0} in
its HW spec.

Signed-off-by: Neo Jou <neojou@gmail.com>